pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gbc"
version = "0.1.0"
description = "Captioning pipeline and data utilities for graph-based captioning."
readme = "README.md"
license = { file="LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apple Sample Code License",
]
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "pyarrow",
    "pydantic>=2.6.4",
    "tqdm>=4.66.2",
    "numpy<2.0.0",
    "transformers>=4.46.3",
    "Pillow>=10.2.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.2",
]

[project.optional-dependencies]
tests = [
    "matplotlib",
    "objprint>=0.2.3",
    "opencv_python>=4.8.0.76",
    "opencv_python_headless>=4.9.0.80",
]
processing = [
    "torch>=2.2.2",
    "nltk",
    "detoxify>=0.5.2",
    "open_clip_torch>=2.24.0",
]
t2i = [
    "torch>=2.5.0",  # >= 2.5.0 for flex attention
    "torchvision",
    "lightning",
    "einops",
    "diffusers>=0.30.0",
    "supervision",  # For bbox annotation
    "scikit-image",      # for segmentation
    "sentencepiece",     # for tokenizer
    "accelerate",
]
captioning = [
    "torch>=2.2.2",
    "torchvision>=0.17.2",
    "lightning>=2.2.2",
    "vllm",
    "hbutils>=0.9.3",
    "sentence_transformers>=2.5.1",
    "scipy>=1.11.4",
    "supervision",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true
